refactor(ReservationRepository): 패키지 이름 변경 및 인터페이스 추가
