+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,0 +1,13 @@
+package roomescape.repository;
+
+import java.util.List;
+import roomescape.domain.Reservation;
+
+public interface ReservationRepository {
+
+    Reservation saveReservation(Reservation reservation);
+
+    List<Reservation> findAllReservation();
+
+    void deleteReservationById(long reservationId);
+}
